pyproject.toml
[project]
name = "exec-assistant"
version = "0.1.0"
description = "Executive Assistant multi-agent system for healthcare infrastructure leadership"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Executive Assistant Team"}
]
dependencies = [
    # Strands Agent SDK
    # "strands-sdk>=0.1.0",  # Temporarily commented for testing

    # AWS SDK and services
    "boto3>=1.34.0",
    "botocore>=1.34.0",

    # Slack integration
    "slack-sdk>=3.27.0",
    "slack-bolt>=1.18.0",

    # Calendar APIs
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",

    # Microsoft Graph (alternative to Google Calendar)
    "msgraph-sdk>=1.0.0",

    # Communication channels
    "twilio>=8.10.0",
    "sendgrid>=6.11.0",

    # Data validation and serialization
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Configuration
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # Utilities
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",

    # Authentication
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "moto[all]>=4.2.0",  # AWS service mocking

    # Linting and formatting
    "ruff>=0.1.0",
    "mypy>=1.7.0",

    # Type stubs
    "boto3-stubs[essential]>=1.34.0",
    "types-pyyaml>=6.0.12",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.19",

    # Pre-commit hooks
    "pre-commit>=3.5.0",
]

infra = [
    # Pulumi for infrastructure as code
    "pulumi>=3.95.0",
    "pulumi-aws>=6.13.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "ARG001", # unused function argument (common in interfaces)
    "ARG002", # unused method argument (common in interfaces)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]  # Allow assert in tests

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "strands.*",
    "moto.*",
    "google.*",
    "msgraph.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=exec_assistant",
    "--cov-report=term-missing",
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require AWS credentials)",
    "e2e: End-to-end tests (require full environment)",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
